Add demo-py for PyPI.

+++ src/demo-lib/src/lib.rs
@@ -8,6 +8,6 @@ mod tests {
     }
 }
 
-pub fn add(a: i64, b: i64) -> i64 {
+pub fn add(a: isize, b: isize) -> isize {
     a + b
 }
